Move classes commonmark-test-util to org.commonmark.testutil package
This is in preparation for Java 9, so that we have one common prefix for
this module.
