+++ commonmark-ext-ins/src/test/java/org/commonmark/ext/ins/InsTest.java
@@ -1,10 +1,10 @@
 package org.commonmark.ext.ins;
 
 import org.commonmark.Extension;
-import org.commonmark.renderer.html.HtmlRenderer;
 import org.commonmark.node.Node;
 import org.commonmark.parser.Parser;
-import org.commonmark.test.RenderingTestCase;
+import org.commonmark.renderer.html.HtmlRenderer;
+import org.commonmark.testutil.RenderingTestCase;
 import org.junit.Test;
 
 import java.util.Collections;
