+++ commonmark-test-util/src/main/java/org/commonmark/testutil/SpecTestCase.java
@@ -1,7 +1,7 @@
-package org.commonmark.test;
+package org.commonmark.testutil;
 
-import org.commonmark.spec.SpecExample;
-import org.commonmark.spec.SpecReader;
+import org.commonmark.testutil.spec.SpecExample;
+import org.commonmark.testutil.spec.SpecReader;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
