+++ commonmark-test-util/src/main/java/org/commonmark/testutil/spec/SpecExample.java
@@ -1,4 +1,4 @@
-package org.commonmark.spec;
+package org.commonmark.testutil.spec;
 
 public class SpecExample {
