+++ commonmark-test-util/src/main/java/org/commonmark/testutil/spec/SpecReader.java
@@ -1,4 +1,4 @@
-package org.commonmark.spec;
+package org.commonmark.testutil.spec;
 
 import java.io.BufferedReader;
 import java.io.IOException;
