+++ commonmark/src/test/java/org/commonmark/test/HtmlRendererTest.java
@@ -7,7 +7,7 @@ import org.commonmark.node.Node;
 import org.commonmark.parser.Parser;
 import org.commonmark.renderer.NodeRenderer;
 import org.commonmark.renderer.html.*;
-import org.commonmark.spec.SpecReader;
+import org.commonmark.testutil.spec.SpecReader;
 import org.junit.Test;
 
 import java.util.*;
