+++ commonmark/src/test/java/org/commonmark/test/SpecBenchmark.java
@@ -2,7 +2,7 @@ package org.commonmark.test;
 
 import org.commonmark.renderer.html.HtmlRenderer;
 import org.commonmark.parser.Parser;
-import org.commonmark.spec.SpecReader;
+import org.commonmark.testutil.spec.SpecReader;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.Scope;
 import org.openjdk.jmh.annotations.State;
